+++ tests/src/rules/no-default-export.js
@@ -1,4 +1,4 @@
-import { test } from '../utils';
+import { test, testVersion } from '../utils';
 
 import { RuleTester } from 'eslint';
 
@@ -85,38 +85,82 @@ ruleTester.run('no-default-export', rule, {
       parser: require.resolve('babel-eslint'),
     }),
   ],
-  invalid: [
-    test({
+  invalid: [].concat(
+    testVersion('> 2', () => ({
       code: 'export default function bar() {};',
-      errors: [{
-        type: 'ExportDefaultDeclaration',
-        message: 'Prefer named exports.',
-      }],
-    }),
-    test({
+      errors: [
+        {
+          type: 'ExportDefaultDeclaration',
+          message: 'Prefer named exports.',
+          line: 1,
+          column: 8,
+        },
+      ],
+    })),
+    testVersion('> 2', () => ({
       code: `
         export const foo = 'foo';
         export default bar;`,
-      errors: [{
-        type: 'ExportDefaultDeclaration',
-        message: 'Prefer named exports.',
-      }],
-    }),
+      errors: [
+        {
+          type: 'ExportDefaultDeclaration',
+          message: 'Prefer named exports.',
+          line: 3,
+          column: 16,
+        },
+      ],
+    })),
+    testVersion('> 2', () => ({
+      code: 'export default class Bar {};',
+      errors: [
+        {
+          type: 'ExportDefaultDeclaration',
+          message: 'Prefer named exports.',
+          line: 1,
+          column: 8,
+        },
+      ],
+    })),
+    testVersion('> 2', () => ({
+      code: 'export default function() {};',
+      errors: [
+        {
+          type: 'ExportDefaultDeclaration',
+          message: 'Prefer named exports.',
+          line: 1,
+          column: 8,
+        },
+      ],
+    })),
+    testVersion('> 2', () => ({
+      code: 'export default class {};',
+      errors: [
+        {
+          type: 'ExportDefaultDeclaration',
+          message: 'Prefer named exports.',
+          line: 1,
+          column: 8,
+        },
+      ],
+    })),
     test({
       code: 'let foo; export { foo as default }',
-      errors: [{
-        type: 'ExportNamedDeclaration',
-        message: 'Do not alias `foo` as `default`. Just export `foo` itself ' +
-          'instead.',
-      }],
+      errors: [
+        {
+          type: 'ExportNamedDeclaration',
+          message: 'Do not alias `foo` as `default`. Just export `foo` itself instead.',
+        },
+      ],
     }),
     test({
       code: 'export default from "foo.js"',
       parser: require.resolve('babel-eslint'),
-      errors: [{
-        type: 'ExportNamedDeclaration',
-        message: 'Prefer named exports.',
-      }],
+      errors: [
+        {
+          type: 'ExportNamedDeclaration',
+          message: 'Prefer named exports.',
+        },
+      ],
     }),
-  ],
+  ),
 });